fixing up a handresult class
